ia-assistente/database.py
# conexão com o banco de dados

import mysql.connector
from mysql.connector import Error
from dotenv import load_dotenv
import os

load_dotenv()

def create_connection():
    try:
        connection = mysql.connector.connect(
            host=os.getenv("MYSQL_HOST", "localhost"),
            database=os.getenv("MYSQL_DATABASE", "startedu"),
            user=os.getenv("MYSQL_USER", "root"),
            password=os.getenv("MYSQL_PASSWORD", "")
        )
        if connection.is_connected():
            return connection
    except Error as e:
        print(f"Erro na conexão com o MySQL: {e}")
        return None
