ia-assistente/requirements.txt
#dependencias

fastapi
uvicorn
langchain
openai
mysql-connector-python  # Se for MySQL
python-dotenv
